# APP03022-mrch-bff-somx-ppsomx-util/cloud-endpoint/openapi.yaml
swagger: '2.0'
info:
  contact:
    email: jrcastillo@falabella.cl
    name: Arquitectura Falabella Tecnologia Corp.
  description: |
    Support RESTful operations for utils flow. This API provides access to parameters, modules, messages, application messages, processes, item types, items, shared catalogs, suppliers, supplier blocks, status train, conversions and catalog layout validation.
  title: Utils API
  version: 2.0.0
  x-business-acronym: mrch
  x-capability-acronym: utils
  x-consumer-segments: retail
  x-service-name: utils
host: '${DOMAIN_OPENAPI}'
schemes:
  - https
paths:
...
  /suppliers:
    get:
      tags:
        - suppliers
      summary: List suppliers (active by default)
      operationId: listSuppliers
      produces:
        - application/json
      parameters:
        - name: status
          in: query
          required: false
          type: integer
          format: int32
      responses:
        '200':
          description: List of suppliers
          schema:
            type: array
            items:
              $ref: '#/definitions/SupplierDto'
      security:
        - bearerAuth: []
    post:
      tags:
        - suppliers
      summary: Create supplier
      operationId: createSupplier
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: X-User-Id
          in: header
          required: false
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SupplierCreateDto'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/SupplierDto'
        '400':
          description: Invalid data
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
  '/suppliers/{id}':
    get:
      tags:
        - suppliers
      summary: Get supplier by ID
      operationId: getSupplierById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Supplier
          schema:
            $ref: '#/definitions/SupplierDto'
        '404':
          description: Not found
      security:
        - bearerAuth: []
    put:
      tags:
        - suppliers
      summary: Update supplier
      operationId: updateSupplier
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
        - name: X-User-Id
          in: header
          required: false
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/SupplierUpdateDto'
      responses:
        '200':
          description: Updated
          schema:
            $ref: '#/definitions/SupplierDto'
        '404':
          description: Not found
      security:
        - bearerAuth: []
    delete:
      tags:
        - suppliers
      summary: Deactivate supplier
      operationId: deleteSupplier
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
        - name: X-User-Id
          in: header
          required: false
          type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
      security:
        - bearerAuth: []
  '/suppliers/number/{supplierNumber}':
    get:
      tags:
        - suppliers
      summary: Get supplier by number
      operationId: getSupplierByNumber
      produces:
        - application/json
      parameters:
        - name: supplierNumber
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Supplier
          schema:
            $ref: '#/definitions/SupplierDto'
        '404':
          description: Not found
      security:
        - bearerAuth: []
  '/suppliers/rfc/{rfc}':
    get:
      tags:
        - suppliers
      summary: Get supplier by RFC
      operationId: getSupplierByRfc
      produces:
        - application/json
      parameters:
        - name: rfc
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Supplier
          schema:
            $ref: '#/definitions/SupplierDto'
        '404':
          description: Not found
      security:
        - bearerAuth: []
...
  /suppliers/filter:
    get:
      tags:
        - suppliers
      summary: 'STM-831: Filter suppliers by type and block status'
      operationId: filterSuppliers
      produces:
        - application/json
      parameters:
        - name: tipoProveedor
          in: query
          required: true
          type: integer
          format: int32
        - name: estatusBloqueo
          in: query
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Filtered list
          schema:
            type: array
            items:
              $ref: '#/definitions/SupplierFilterDto'
      security:
        - bearerAuth: []
...
  '/security/user-attributes-by-key/{userKey}':
    get:
      tags:
        - security
      summary: Get user attributes (ATR001/ATR002/ATR004) by user_data sub key
      description: STM-1525 — consumido por fiscal-api y finanzas-api para resolver contexto de seguridad desde el JWT sub
      operationId: getUserAttributesByKey
      produces:
        - application/json
      parameters:
        - name: userKey
          in: path
          required: true
          description: User lookup key (sub claim from JWT)
          type: string
      responses:
        '200':
          description: User attributes
          schema:
            type: object
            properties:
              success:
                type: boolean
              data:
                type: object
                properties:
                  userDataId:
                    type: integer
                  sub:
                    type: string
                  preferredUsername:
                    type: string
                  email:
                    type: string
                  attributes:
                    type: array
                    items:
                      type: object
                      properties:
                        typeKey:
                          type: string
                        valueKey:
                          type: string
        '400':
          description: Bad request — userKey obligatorio
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '404':
          description: User not found / inactive
          schema:
            $ref: '#/definitions/ExceptionWrapper'
        '500':
          description: Internal error
          schema:
            $ref: '#/definitions/ExceptionWrapper'
      security:
        - bearerAuth: []
...
definitions:
  ExceptionWrapper:
    type: object
    properties:
      code:
        description: Numeric HTTP Status error
        example: 401
        format: int32
        type: integer
      message:
        description: Error description
        example: Unauthorized
        type: string
...
  SupplierTypeDto:
    type: object
    properties:
      id:
        type: integer
        format: int32
      code:
        type: string
      description:
        type: string
  PaymentConditionDto:
    type: object
    properties:
      id:
        type: integer
        format: int32
      conditionName:
        type: string
      days:
        type: integer
        format: int32
  SupplierDto:
    type: object
    properties:
      id:
        type: integer
        format: int32
      supplierNumber:
        type: string
      rfc:
        type: string
      businessName:
        type: string
      supplierType:
        $ref: '#/definitions/SupplierTypeDto'
      logo:
        type: string
      paymentCondition:
        $ref: '#/definitions/PaymentConditionDto'
      emailPrincipal:
        type: string
        description: Email Principal del proveedor.
      emailFinancial:
        type: string
        description: Email de Cuentas por Cobrar del proveedor.
      emailCommercial:
        type: string
        description: Email de la cuenta clave comercial del proveedor.
      status:
        type: integer
        format: int32
  SupplierCreateDto:
    type: object
    required:
      - supplierNumber
      - rfc
      - businessName
      - emailPrincipal
      - emailFinancial
    properties:
      supplierNumber:
        type: string
        example: PROV001
      rfc:
        type: string
        example: ABC123456789
      businessName:
        type: string
      supplierTypeId:
        type: integer
        format: int32
      logo:
        type: string
      paymentConditionId:
        type: integer
        format: int32
      emailPrincipal:
        type: string
        description: Email Principal del proveedor (obligatorio).
        example: principal@proveedor.com
      emailFinancial:
        type: string
        description: Email de Cuentas por Cobrar (obligatorio).
        example: cxc@proveedor.com
      emailCommercial:
        type: string
        description: Email de la cuenta clave comercial (opcional).
        example: comercial@proveedor.com
  SupplierUpdateDto:
    type: object
    properties:
      rfc:
        type: string
      businessName:
        type: string
      supplierTypeId:
        type: integer
        format: int32
      logo:
        type: string
      paymentConditionId:
        type: integer
        format: int32
      emailPrincipal:
        type: string
        description: Email Principal del proveedor.
      emailFinancial:
        type: string
        description: Email de Cuentas por Cobrar.
      emailCommercial:
        type: string
        description: Email de la cuenta clave comercial.
      status:
        type: integer
        format: int32
  SupplierBlockInfoDto:
    type: object
    properties:
      validFrom:
        type: string
        format: date
      validTo:
        type: string
        format: date
      blockReason:
        type: string
  SupplierFilterDto:
    type: object
    properties:
      id:
        type: integer
        format: int32
      supplierNumber:
        type: string
      rfc:
        type: string
      businessName:
        type: string
      supplierType:
        $ref: '#/definitions/SupplierTypeDto'
      paymentCondition:
        $ref: '#/definitions/PaymentConditionDto'
      status:
        type: integer
        format: int32
      blocked:
        type: boolean
      blockInfo:
        $ref: '#/definitions/SupplierBlockInfoDto'
...
securityDefinitions:
  api_key:
    in: query
    name: key
    type: apiKey
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
  keycloak:
    authorizationUrl: ''
    flow: implicit
    scopes: {}
    type: oauth2
    x-google-audiences: vendor-backend
    x-google-issuer: '${KEYCLOAK}/auth/realms/corp'
    x-google-jwks_uri: '${JWKS_URL}'
    x-google-jwt-locations:
      - header: Authorization
        value_prefix: 'Bearer '
security:
  - bearerAuth: []
...